requirements.txt
python-dotenv
ollama
python-docx
whisper
PyMuPDF
